+++ packages/access-definitions/src/Wrapping/Contracts/TypeRetrievalAccessException.php
@@ -3,67 +3,17 @@
 declare(strict_types=1);
 
 namespace EDT\Wrapping\Contracts;
-use function implode;
 
 class TypeRetrievalAccessException extends AccessException
 {
     /**
-     * @param non-empty-string $typeIdentifier
-     * @param list<non-empty-string> $availableTypeIdentifiers
+     * @param non-empty-string       $typeIdentifier
+     * @param list<non-empty-string> $reasons
      */
-    public static function unknownTypeIdentifier(string $typeIdentifier, array $availableTypeIdentifiers): self
+    public static function notPresent(string $typeIdentifier, array $reasons): self
     {
-        $typeIdentifiersString = implode(', ', $availableTypeIdentifiers);
-        $self = new self("Type identifier '$typeIdentifier' not known. Known type identifiers are: $typeIdentifiersString.");
-        $self->typeIdentifier = $typeIdentifier;
+        $reasonsString = implode(' ', $reasons);
 
-        return $self;
-    }
-
-    /**
-     * @template TType of \EDT\Wrapping\Contracts\Types\TypeInterface
-     *
-     * @param non-empty-string $typeIdentifier
-     * @param class-string<TType> $implementation
-     */
-    public static function noNameWithImplementation(string $typeIdentifier, string $implementation): self
-    {
-        $self = new self("Type with identifier '$typeIdentifier' exists but does not implement $implementation.");
-        $self->typeIdentifier = $typeIdentifier;
-
-        return $self;
-    }
-
-    /**
-     * @param non-empty-string $typeIdentifier
-     */
-    public static function typeExistsButNotAvailable(string $typeIdentifier): self
-    {
-        $self = new self("Type '$typeIdentifier' exists but is not available.");
-        $self->typeIdentifier = $typeIdentifier;
-
-        return $self;
-    }
-
-    /**
-     * @param non-empty-string $typeIdentifier
-     */
-    public static function typeExistsButNotDirectlyAccessible(string $typeIdentifier): self
-    {
-        $self = new self("Type '$typeIdentifier' exists but is not directly accessible.");
-        $self->typeIdentifier = $typeIdentifier;
-
-        return $self;
-    }
-
-    /**
-     * @param non-empty-string $typeIdentifier
-     */
-    public static function typeExistsButNotReferencable(string $typeIdentifier): self
-    {
-        $self = new self("Type '$typeIdentifier' exists but is not referencable.");
-        $self->typeIdentifier = $typeIdentifier;
-
-        return $self;
+        return new self("Type instance with identifier '$typeIdentifier' matching the defined criteria was not found due to the following reasons: $reasonsString");
     }
 }